+++ CHANGELOG.md
@@ -2,13 +2,28 @@
 # Change Log
 All notable changes to this project will be documented in this file.
  
-## [ 0.2.0 ] - Unreleased
+## [ 0.3.0 ] - Unreleased
  
 ### Added
 
 ### Changed
  
 ### Fixed
+ 
+## [ 0.2.0 ] - 2022-12-23
+ 
+### Added
+
+- (fdr) Palmgren-miner damage rule
+- (lcc) Rychlik counting method
+- (lsg) Autoregressive model
+- (lsg) Random walk
+- (utils) Cycle counting aggregation
+- (utils) Fitter for SN curve
+
+### Changed
+
+- (utils) `getSequencePeakAndValleys` was changed to `sequencePeakAndValleys`
  
 ## [ 0.1.0 ] - 2022-12-11